+++ src/InfluxDB2/Client.php
@@ -42,11 +42,12 @@ class Client
      *          'flushInterval' => flush data at least in this interval
      *      ]
      * @param array|null $writeOptions Array containing the write parameters (See above)
+     * @param array|null $pointSettings
      * @return WriteApi
      */
-    public function createWriteApi(array $writeOptions = null): WriteApi
+    public function createWriteApi(array $writeOptions = null, array $pointSettings = null): WriteApi
     {
-        $writeApi = new WriteApi($this->options, $writeOptions);
+        $writeApi = new WriteApi($this->options, $writeOptions, $pointSettings);
         $this->autoCloseable[] = $writeApi;
         return $writeApi;
     }