+++ src/InfluxDB2/WriteApi.php
@@ -38,6 +38,42 @@ class WriteOptions
     }
 }
 
+class PointSettings
+{
+    private $defaultTags;
+
+    public function __construct(array $defaultTags = null)
+    {
+        $this->defaultTags = array();
+
+        if (!empty($defaultTags)) {
+            foreach (array_keys($defaultTags) as $key) {
+                $this->addDefaultTag($key, $defaultTags[$key]);
+            }
+        }
+    }
+
+    public function addDefaultTag(string $key, string $value)
+    {
+        $this->defaultTags[$key] = $this->getValue($value);
+    }
+
+    private function getValue(string $value): string
+    {
+        if (substr( $value, 0, 6 ) === '${env.')
+        {
+            return getenv(substr( $value, 6, strlen($value) - 7));
+        }
+
+        return $value;
+    }
+
+    public function getDefaultTags()
+    {
+        return $this->defaultTags;
+    }
+}
+
 /**
  * Write time series data into InfluxDB.
  * @package InfluxDB2
@@ -45,6 +81,7 @@ class WriteOptions
 class WriteApi extends DefaultApi
 {
     public $writeOptions;
+    public $pointSettings;
 
     /** @var Worker */
     private $worker;
@@ -53,12 +90,14 @@ class WriteApi extends DefaultApi
     /**
      * WriteApi constructor.
      * @param $options
-     * @param $writeOptions
+     * @param array $writeOptions
+     * @param array|null $pointSettings
      */
-    public function __construct($options,array $writeOptions = null)
+    public function __construct($options, array $writeOptions = null, array $pointSettings = null)
     {
         parent::__construct($options);
         $this->writeOptions = new WriteOptions($writeOptions) ?: new WriteOptions();
+        $this->pointSettings = new PointSettings($pointSettings) ?: new PointSettings();
     }
 
     /**
@@ -99,6 +138,8 @@ class WriteApi extends DefaultApi
         $this->check("bucket", $bucketParam);
         $this->check("org", $orgParam);
 
+        $this->addDefaultTags($data);
+
         $payload = $this->generatePayload($data, $precisionParam, $bucketParam, $orgParam);
 
         if ($payload == null) {
@@ -110,6 +151,36 @@ class WriteApi extends DefaultApi
             $this->worker()->push($payload);
         } else {
             $this->writeRaw($payload, $precisionParam, $bucketParam, $orgParam);
+        }
+    }
+
+    private function addDefaultTags(&$data)
+    {
+        $defaultTags = $this->pointSettings->getDefaultTags();
+
+        if (is_array($data))
+        {
+            if (array_key_exists('name', $data))
+            {
+                foreach (array_keys($defaultTags) as $key)
+                {
+                    $data['tags'][$key] = $defaultTags[$key];
+                }
+            }
+            else
+            {
+                foreach ($data as &$item)
+                {
+                    $this->addDefaultTags($item);
+                }
+            }
+        }
+        elseif ($data instanceof Point)
+        {
+            foreach (array_keys($defaultTags) as $key)
+            {
+                $data->addTag($key, $defaultTags[$key]);
+            }
         }
     }